Move ServiceNamespacesContext its own class
